Added NoDB and added clear() method in serachbar
This is a mini-joke as I was going to implement everything with mongo but this approach is better, nuxt-content is far way easier to use and manage also, no need for servers and exporting it with capacitor it's also better. I also made JQL  based on jslinq to manage Equreka in the near-future so much work for this but you know this is how dev-ing work HAHAHA. Anyway getting late, bye.